+++ Ejercicio11.java
@@ -11,9 +11,9 @@ import java.util.Scanner;
 public class Ejercicio11 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        float num;
+        double num;
         System.out.print("Introduce la cantidad en Kb: ");
-        num = sc.nextInt();
+        num = sc.nextDouble();
         System.out
                 .print(String.format("%.0f", (num)) + " Kb equivalen a " + String.format("%.2f", (num / 1024)) + " Mb");
         sc.close();
